fix: fix requests

+++ src/routes/commentsRouter.ts
@@ -4,6 +4,7 @@ import { inputValidationMiddleware } from '../middlewares';
 import { authMiddleware } from '../middlewares/authMiddleware';
 import { commentsService } from '../domain/comments-service';
 import { body } from 'express-validator';
+import { usersQueryRepository } from '../repositories/usersQueryRepository';
 
 export const commentsRouter = Router({});
 
@@ -21,9 +22,9 @@ commentsRouter
     inputValidationMiddleware,
     async (req: Request, res: Response) => {
       const { id } = req.params;
-
+      const currentUser = await usersQueryRepository.getUserById(req.user.id);
       const comment = await commentsQueryRepository.getCommentById(id);
-      if (comment) {
+      if (comment && currentUser && comment.userId === currentUser.id) {
         const isDeleted = await commentsService.deleteCommentById(id);
         isDeleted && res.send(204);
         return;
@@ -48,8 +49,9 @@ commentsRouter
       const { id } = req.params;
       const { content } = req.body;
       const comment = await commentsQueryRepository.getCommentById(id);
+      const currentUser = await usersQueryRepository.getUserById(req.user.id);
 
-      if (comment) {
+      if (comment && currentUser && comment.userId === currentUser.id) {
         const isUpdated = await commentsService.updateCommentById(id, content);
         isUpdated && res.send(204);
         return;
